prevent emacsclient accessing org-mode too soon
I am displaying output of this function on my system panel.
It is executed every 10 seconds by emacs client.

Sometimes when I launch emacs, this panel / emacsclient
tries to load org-pomodor before the whole org-mode is
loaded by doom's magic. And that causes weird issues with
keybinding and a lot of weird messages.

This commit prevents execution of this function when
org-mode is not loaded (yet)