treat certain buffers as popups while switching
when switching buffers via default ivy interface, check if buffer
belongs to "help buffers" - which are buffers opened for quick search
of information and reference - and in such case pop them instead of
switch to them. With popup rules in place, help buffers will always
appear in predictable space and will be easy to close them and
manipulate them as other popup buffers.